+++ gradle.properties
@@ -1,6 +1,6 @@
 # IntelliJ Platform Artifacts Repositories -> https://plugins.jetbrains.com/docs/intellij/intellij-artifacts.html
 
-pluginGroup = com.semgrep.semgrep
+pluginGroup = com.semgrep.idea
 pluginName = semgrep
 pluginRepositoryUrl = https://github.com/returntocorp/semgrep-intellij
 # SemVer format -> https://semver.org
