+++ src/main/kotlin/com/semgrep/idea/lsp/SemgrepLanguageServer.kt
@@ -1,8 +1,8 @@
-package com.semgrep.semgrep.lsp
+package com.semgrep.idea.lsp
 
-import com.semgrep.semgrep.lsp.custom_notifications.ScanWorkspaceParams
-import com.semgrep.semgrep.lsp.custom_requests.LoginResult
-import com.semgrep.semgrep.lsp.custom_requests.LoginStatusResult
+import com.semgrep.idea.lsp.custom_notifications.ScanWorkspaceParams
+import com.semgrep.idea.lsp.custom_requests.LoginResult
+import com.semgrep.idea.lsp.custom_requests.LoginStatusResult
 import org.eclipse.lsp4j.jsonrpc.services.JsonNotification
 import org.eclipse.lsp4j.jsonrpc.services.JsonRequest
 import org.eclipse.lsp4j.services.LanguageServer
