+++ src/main/kotlin/com/semgrep/idea/lsp/actions/LoginAction.kt
@@ -1,14 +1,13 @@
-package com.semgrep.semgrep.lsp.actions
+package com.semgrep.idea.lsp.actions
 
 import com.intellij.ide.BrowserUtil
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.platform.lsp.api.LspServer
-import com.semgrep.semgrep.lsp.SemgrepLspServer
-import com.semgrep.semgrep.lsp.custom_notifications.LoginFinishRequest
-import com.semgrep.semgrep.lsp.custom_requests.LoginRequest
+import com.semgrep.idea.lsp.custom_notifications.LoginFinishRequest
+import com.semgrep.idea.lsp.custom_requests.LoginRequest
 
 class LoginAction: LspAction() {
-    override fun actionPerformed(e: AnActionEvent, servers: List<SemgrepLspServer>) {
+    override fun actionPerformed(e: AnActionEvent, servers: List<com.semgrep.idea.lsp.SemgrepLspServer>) {
         val loginRequest = LoginRequest(servers.first())
         val response = (servers.first() as LspServer).requestExecutor.sendRequestSync(loginRequest)?: return
         BrowserUtil.browse(response.url)
