+++ src/main/kotlin/com/semgrep/idea/lsp/actions/LogoutAction.kt
@@ -1,11 +1,10 @@
-package com.semgrep.semgrep.lsp.actions
+package com.semgrep.idea.lsp.actions
 
 import com.intellij.openapi.actionSystem.AnActionEvent
-import com.semgrep.semgrep.lsp.SemgrepLspServer
-import com.semgrep.semgrep.lsp.custom_notifications.LogoutNotifcation
+import com.semgrep.idea.lsp.custom_notifications.LogoutNotifcation
 
 class LogoutAction:LspAction(){
-    override fun actionPerformed(e: AnActionEvent, servers: List<SemgrepLspServer>) {
+    override fun actionPerformed(e: AnActionEvent, servers: List<com.semgrep.idea.lsp.SemgrepLspServer>) {
         val logoutNotifcation = LogoutNotifcation(servers.first())
         servers.map {
             it.requestExecutor.sendNotification(LogoutNotifcation(it))
