+++ src/main/kotlin/com/semgrep/idea/lsp/actions/LspAction.kt
@@ -1,17 +1,16 @@
-package com.semgrep.semgrep.lsp.actions
+package com.semgrep.idea.lsp.actions
 
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
-import com.semgrep.semgrep.lsp.SemgrepLspServer
 
 abstract class LspAction:AnAction() {
     override fun actionPerformed(e: AnActionEvent) {
         val project = e.project ?: return
-        val servers = SemgrepLspServer.getInstances(project)
+        val servers = com.semgrep.idea.lsp.SemgrepLspServer.getInstances(project)
         if (servers.isEmpty()){
             return
         }
         actionPerformed(e, servers)
     }
-    abstract fun actionPerformed(e: AnActionEvent, servers: List<SemgrepLspServer>);
+    abstract fun actionPerformed(e: AnActionEvent, servers: List<com.semgrep.idea.lsp.SemgrepLspServer>);
 }
