+++ src/main/kotlin/com/semgrep/idea/lsp/custom_notifications/LogoutNotifcation.kt
@@ -1,0 +1,10 @@
+package com.semgrep.idea.lsp.custom_notifications
+
+import com.intellij.platform.lsp.api.requests.LspClientNotification
+
+class LogoutNotifcation(private val server: com.semgrep.idea.lsp.SemgrepLspServer): LspClientNotification(server){
+    override fun sendNotification() {
+        server.lsp4jServer.logout()
+    }
+
+}
