+++ src/main/kotlin/com/semgrep/idea/lsp/custom_requests/LoginStatusRequest.kt
@@ -1,12 +1,11 @@
-package com.semgrep.semgrep.lsp.custom_requests
+package com.semgrep.idea.lsp.custom_requests
 
 import com.intellij.platform.lsp.api.requests.LspRequest
-import com.semgrep.semgrep.lsp.SemgrepLspServer
 import java.util.concurrent.CompletableFuture
 
 data class LoginStatusResult(val loggedIn: Boolean)
 
-class LoginStatusRequest(private val server: SemgrepLspServer): LspRequest<LoginStatusResult,LoginStatusResult>(server) {
+class LoginStatusRequest(private val server: com.semgrep.idea.lsp.SemgrepLspServer): LspRequest<LoginStatusResult,LoginStatusResult>(server) {
     override fun preprocessResponse(serverResponse: LoginStatusResult): LoginStatusResult {
         return serverResponse
     }
