+++ src/main/kotlin/com/semgrep/idea/lsp/settings/SemgrepLspServerSettings.kt
@@ -1,4 +1,4 @@
-package com.semgrep.semgrep.lsp.settings
+package com.semgrep.idea.lsp.settings
 
 enum class TraceLevel {
     OFF,
