--- src/main/kotlin/com/semgrep/semgrep/lsp/custom_notifications/RefreshRulesNotification.kt
@@ -1,10 +1,0 @@
-package com.semgrep.semgrep.lsp.custom_notifications
-
-import com.intellij.platform.lsp.api.requests.LspClientNotification
-import com.semgrep.semgrep.lsp.SemgrepLspServer
-
-class RefreshRulesNotification(private val server: SemgrepLspServer): LspClientNotification(server){
-    override fun sendNotification() {
-        server.lsp4jServer.refreshRules()
-    }
-}
