--- src/main/kotlin/com/semgrep/semgrep/lsp/custom_notifications/ScanWorkspaceNotification.kt
@@ -1,11 +1,0 @@
-package com.semgrep.semgrep.lsp.custom_notifications
-
-import com.intellij.platform.lsp.api.requests.LspClientNotification
-import com.semgrep.semgrep.lsp.SemgrepLspServer
-
-data class ScanWorkspaceParams(val full: Boolean)
-class ScanWorkspaceNotification(private val server: SemgrepLspServer, private val params: ScanWorkspaceParams): LspClientNotification(server){
-    override fun sendNotification() {
-        server.lsp4jServer.scanWorkspace(params)
-    }
-}
