+++ src/test/kotlin/com/semgrep/idea/MyPluginTest.kt
@@ -1,0 +1,7 @@
+package com.semgrep.idea
+
+import com.intellij.testFramework.TestDataPath
+import com.intellij.testFramework.fixtures.BasePlatformTestCase
+@TestDataPath("\$CONTENT_ROOT/src/test/testData")
+class MyPluginTest : BasePlatformTestCase() {
+}
